+++ composer/datasets/mnist_hparams.py
@@ -1,7 +1,7 @@
 # Copyright 2022 MosaicML Composer authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""MNIST image classification dataset.
+"""MNIST image classification dataset hyperparameters.
 
 The MNIST dataset is a collection of labeled 28x28 black and white images of handwritten examples of the numbers 0-9.
 See the `wikipedia entry <https://en.wikipedia.org/wiki/MNIST_database>`_ for more details.
@@ -12,9 +12,9 @@ from dataclasses import dataclass
 import yahp as hp
 from torchvision import datasets, transforms
 
-from composer.datasets.dataloader import DataLoaderHparams
-from composer.datasets.hparams import DatasetHparams, SyntheticHparamsMixin
+from composer.datasets.dataset_hparams import DataLoaderHparams, DatasetHparams
 from composer.datasets.synthetic import SyntheticBatchPairDataset
+from composer.datasets.synthetic_hparams import SyntheticHparamsMixin
 from composer.utils import dist
 
 __all__ = ["MNISTDatasetHparams"]
