+++ tests/fixtures/dummy_fixtures.py
@@ -10,11 +10,12 @@ from torch.optim import Optimizer
 
 from composer.core import Precision, State
 from composer.core.types import PyTorchScheduler
-from composer.datasets import DataLoaderHparams, DatasetHparams
+from composer.datasets.dataset_hparams import DataLoaderHparams, DatasetHparams
+from composer.datasets.dataset_hparams_registry import dataset_registry
 from composer.models import ModelHparams
 from composer.optim import ExponentialScheduler
 from composer.optim.optimizer_hparams_registry import AdamHparams
-from composer.trainer.trainer_hparams import TrainerHparams, dataset_registry, model_registry
+from composer.trainer.trainer_hparams import TrainerHparams, model_registry
 from tests.common import RandomClassificationDatasetHparams, SimpleModel, SimpleModelHparams
